기존 스레드 로딩 컴포넌트를 이동하고, 새로운 스레드 페이지 구조를 추가하여 쿼리 클라이언트를 통합함. 불필요한 파일 삭제 및 경로 수정.
