+++ app/main/trash/page.tsx
@@ -1,83 +1,53 @@
 "use client";
 
-import { DefaultThreadItem } from "@/components/thread-item";
 import { Button } from "@/components/ui/button";
 import { ThreadItemType } from "@/lib/typeDefs";
 import { createClient } from "@/utils/supabase/client";
+import {
+  QueryClient,
+  QueryClientProvider,
+  useQuery,
+} from "@tanstack/react-query";
 import { AnimatePresence } from "framer-motion";
-import { useEffect, useState } from "react";
+import { DefaultThreadItem } from "../thread-item";
+import { ThreadLoading } from "../thread-loading";
 
-export default function TrashThread() {
-  //TODO: implement this function
+export default function TrashPage() {
+  const queryClient = new QueryClient();
 
-  const supabase = createClient();
-  const [trashThreads, setTrashThreads] = useState<ThreadItemType[] | null>(
-    null
+  return (
+    <QueryClientProvider client={queryClient}>
+      <TrashThread />
+    </QueryClientProvider>
   );
+}
 
-  useEffect(() => {
-    const fetchTrashThreads = async () => {
+function TrashThread() {
+  const supabase = createClient();
+
+  const { data, isLoading, error } = useQuery({
+    queryKey: ["trash"],
+    queryFn: async () => {
       const { data, error } = await supabase
         .from("trash")
         .select()
         .order("created_at", { ascending: false })
         .order("id", { ascending: false });
 
-      console.log("🚀 ~ getTrashThreads ~ data", data);
-
       if (error) {
-        console.error(error);
+        throw new Error(error.message);
       }
 
-      setTrashThreads(data);
-    };
-    fetchTrashThreads();
-  }, [supabase]);
+      return data;
+    },
+  });
 
   const RestoreButton = ({ thread }: { thread: ThreadItemType }) => {
-    const [isMoving, setIsMoving] = useState(false);
-
     return (
       <Button
         size={`sm`}
         onClick={async (e) => {
           e.stopPropagation();
-          setIsMoving(true);
-
-          const { error: DeleteError } = await supabase
-            .from("trash")
-            .delete()
-            .eq("id", thread.id);
-          if (DeleteError) {
-            console.error("🚀 ~ removeThread ~ error", DeleteError);
-            return;
-          }
-
-          const { data, error: moveError } = await supabase
-            .from("new-threads")
-            .insert([
-              {
-                type: thread.type,
-                url: thread.url,
-                title: thread.title,
-                description: thread.description,
-                host: thread.host,
-              },
-            ]);
-
-          if (moveError) {
-            console.error("🚀 ~ removeThread ~ error", moveError);
-            return;
-          }
-
-          //remove from treads
-          setTrashThreads((prev) =>
-            prev
-              ? prev.filter((prevThread) => prevThread.id !== thread.id)
-              : null
-          );
-
-          setIsMoving(false);
         }}
       >
         Restore
@@ -87,22 +57,14 @@ export default function TrashThread() {
 
   return (
     <div className="container mx-auto p-4">
-      <h5 className="self-start m-2">{`Items: ${
-        trashThreads?.length ?? 0
-      }`}</h5>
       <div className="flex flex-col gap-2">
+        {isLoading && <ThreadLoading />}
         <AnimatePresence>
-          {trashThreads?.map((thread) => (
+          {data?.map((thread) => (
             <DefaultThreadItem
               key={thread.id}
               thread={thread}
               threadName="trash"
-              disablePrimaryAction
-              extraButtons={
-                <>
-                  <RestoreButton thread={thread} />
-                </>
-              }
             />
           ))}
         </AnimatePresence>
